refactor: sse hook 분리, evnet 함수화
